Automatically backed up by Learn

+++ lib/cash_register.rb
@@ -14,7 +14,7 @@ class CashRegister
 
 
   def apply_discount
-    @total -= (@total * @discount.to_f / 100).to_f
+    @total -= (@total * @discount.to_f / 100).to_i
     if @discount == 0
       "There is no discount to apply."
     else
